+++ ankusa_kafka/lib/ankusa_kafka.ex
@@ -1,0 +1,18 @@
+defmodule AnkusaKafka do
+  @moduledoc """
+  Documentation for `AnkusaKafka`.
+  """
+
+  @doc """
+  Hello world.
+
+  ## Examples
+
+      iex> AnkusaKafka.hello()
+      :world
+
+  """
+  def hello do
+    :world
+  end
+end