+++ ankusa_kafka/test/ankusa_kafka_test.exs
@@ -1,0 +1,8 @@
+defmodule AnkusaKafkaTest do
+  use ExUnit.Case
+  doctest AnkusaKafka
+
+  test "greets the world" do
+    assert AnkusaKafka.hello() == :world
+  end
+end